refactor: 카테고리 서비스 이름 변경 및 패키지 이동 ( #49 )
